@Peppone98 2025-01-13
## Details of the new implementation
This version introduces several improvements to the computation of forces. The process can be summarized as below
1. **Spatial sorting**: atom indices are spatially sorted based on the Morton code using the `sorted_Morton_seq` function. 
2. **Box edges computation**: the edges of the boxes are computed efficiently with `kernel_min_max!`.
3. **Boolean matrix compression**: the Boolean matrices `sys.neighbor_finder.eligible` and `sys.neighbor_finder.special` are compressed into `UInt32` representations and ordered via the new `compress_boolean_matrices!` kernel. This compression eliminates the overhead associated with reading matrices in the `force_kernel!`. The kernel now accesses only a single element instead of 32.
4. **Force kernel execution**: `force_kernel!` is invoked for 10 iterations consecutively, using the same boxes and matrices stored in the `ForcesBuffer` struct. The number of iterations can be adjusted by setting the `n_steps_reorder` parameter in `GPUNeighborFinder` when defining the system. Setting the `always_inline` flag to `true` in the kernel call significantly improves performance. Below, the new structure of `ForcesBuffer`:
```
struct ForcesBuffer{F, C, M, R}
    fs_mat::F
    box_mins::C
    box_maxs::C
    Morton_seq::M
    compressed_eligible::R
    compressed_special::R
end
```

### Notable Changes to force_kernel!
- **Shared memory optimization**: The kernel now minimises the use of shared memory, employing it solely for storing forces, while maximising the use of fast thread registers.
- **Bit rotation**: elements of the compressed `UInt32` matrices are "rotated" during each interaction evaluation, where the most significant bit becomes the least significant bit. See the example below with `m` in `1:warpsize()`
```
excl = (eligible_bitmask >> (warpsize() - m)) | (eligible_bitmask << m)
```
 
### Note on potential energy calculation 
The calculation of potential energy remains unchanged. Direct access to Boolean matrices via shared memory continues to be faster than the two step process of retrieving the compressed matrices and accessing them within the kernel.


## Standalone force computation
Benchmark of `forces(sys)` performed on `6mrr_equil.pdb` (protein system with 15954 atoms) for the proposed implementation and the previous implementation in Molly: 
|          | `Float32` | `Float64` |
| -------- | -------   | -------   |
| with reordering  | 2.164 ms ± 198.713 μs | 8.109 ms ± 364.257 μs |
| without reordering  | 491.569 μs ±   9.123 μs | 5.867 ms ± 78.764 μs |
| previous implementation (with neighbors calculations)| 15.084 ms ± 399.510 μs |  20.113 ms ± 435.548 |
| previous implementation (without neighbors calculations)| 1.257 ms ± 115.581 μs |  1.585 ms ± 80.793 μs |

## Performances in simulations
Comparison between the performances in simulations of `6mrr_equil.pdb` and a system of 16000 LJ particles placed randomly in a cubic box of 6 nm. The Velocity Verlet algorithm was used, and the reported times correspond to 1000 simulation steps:

|          | `Float32` | `Float64` |
| -------- | -------   | -------   |
| `6mrr_equil.pdb`  |  0.78 s | 6.41 s |
| Random system | 0.82 s | 4.93 s |
| `6mrr_equil.pdb`, previous implementation | 2.82 s |  3.62 s |
| Random system, previous implementation | 2.08 s |  3.19 s |


These tests were condicted on a single `NVIDIA GeForce RTX 4090` GPU. 
As indicated by the GPU specifications at https://www.techpowerup.com/gpu-specs/geforce-rtx-4090.c3889, the theoretical performance is 82.58 TFLOPS for `Float32` operations but only 1,290 GFLOPS (1:64 ratio) for `Float64` operations. This significant disparity reflects the hardware's de-prioritization of `Float64` operations, leading to their markedly slower performance.

### Validation
All tests in the `GROUP=NotGradients` collection passed successfully.
 
### Performance in smaller systems
The advantages of the new algorithm are evident even in smaller systems. For example:
```
n_steps = 1_000
temp = 298.0f0
press = Float32(ustrip(u"u * nm^-1 * ps^-2", 1.0f0u"bar"))
s = System(
    Float32,
    joinpath(data_dir, "5XER", "gmx_coords.gro"),
    joinpath(data_dir, "5XER", "gmx_top_ff.top");
    loggers=(
        temp=TemperatureLogger(Float32, 10),
        coords=CoordinatesLogger(Float32, 10),
        energy=TotalEnergyLogger(Float32, 10),
    ),
    gpu=true,
    units=false,
)
thermostat = AndersenThermostat(temp, 10.0f0)
barostat = MonteCarloBarostat(press, temp, s.boundary; n_steps=20)
simulator = VelocityVerlet(dt=0.002f0, coupling=(thermostat, barostat))

atoms = Array(s.atoms)
s.velocities = CuArray([random_velocity(mass(a), temp) .* 0.01f0 for a in atoms])
simulate!(deepcopy(s), simulator, 100; n_threads=1)
@time simulate!(s, simulator, n_steps; n_threads=1)
```
This simulation of 1000 steps is performed in 0.48 s, while the old implementation takes 1.54 s.